+++ hyperadmin/contrib/apikey/common.py
@@ -1,0 +1,9 @@
+from hyperadmin.contrib.apikey.models import ApiKey
+
+def authenticate(key):
+    try:
+        entry = ApiKey.objects.get(key=key, active=True)
+    except ApiKey.DoesNotExist:
+        return None
+    return entry.user
+