+++ hyperadmin/resources/auth/resources.py
@@ -22,8 +22,11 @@ class AuthResource(BaseResource):
         ])
         return endpoints
     
-    def api_permission_check(self, request):
+    def api_permission_check(self, api_request, endpoint):
         return None #resource is accessible to all
     
     def get_main_link_name(self):
         return 'login'
+    
+    def get_index_endpoint(self):
+        return self.endpoints['login']